requirements.txt
flask
selenium
Pillow
numpy
